+++ sstemplates-core/src/test/java/com/carbonfive/sstemplates/tags/IfTagTest.java
@@ -20,10 +20,10 @@ public class IfTagTest extends TagTestBase
     SsTemplateContext templateContext = renderWorkbook("if.sst");
 
     HSSFRow row = templateContext.getWorkbook().getSheetAt(0).getRow(0);
-    assertNotNull(row.getCell((short) 0));
-    assertEquals("Here I Am 0", row.getCell((short) 0).getStringCellValue());
-    assertNotNull(row.getCell((short) 1));
-    assertEquals("Here I Am 2", row.getCell((short) 1).getStringCellValue());
+    assertNotNull(row.getCell(0));
+    assertEquals("Here I Am 0", row.getCell(0).getStringCellValue());
+    assertNotNull(row.getCell(1));
+    assertEquals("Here I Am 2", row.getCell(1).getStringCellValue());
   }
 
   public void childRenderTest( SsTemplateContext context )
